[feat] create SwipingViewCoordinating to be able handle more Coordinators
